+++ lib-multisrc/machinetranslations/assets/fonts/LICENSE.txt
