+++ lib-multisrc/machinetranslations/build.gradle.kts
@@ -1,0 +1,5 @@
+plugins {
+    id("lib-multisrc")
+}
+
+baseVersionCode = 1
