+++ lib-multisrc/machinetranslations/src/eu/kanade/tachiyomi/multisrc/machinetranslations/MachineTranslationsDto.kt
@@ -1,4 +1,4 @@
-package eu.kanade.tachiyomi.extension.all.snowmtl
+package eu.kanade.tachiyomi.multisrc.machinetranslations
 
 import android.graphics.Color
 import android.os.Build
@@ -8,6 +8,7 @@ import kotlinx.serialization.Serializable
 import kotlinx.serialization.builtins.ListSerializer
 import kotlinx.serialization.json.JsonArray
 import kotlinx.serialization.json.JsonElement
+import kotlinx.serialization.json.JsonObject
 import kotlinx.serialization.json.JsonTransformingSerializer
 import kotlinx.serialization.json.buildJsonObject
 import kotlinx.serialization.json.jsonArray
@@ -31,14 +32,17 @@ data class Dialog(
     val y1: Float,
     val x2: Float,
     val y2: Float,
-    val text: String,
     val angle: Float = 0f,
     val isBold: Boolean = false,
     val isNewApi: Boolean = false,
+    val textByLanguage: Map<String, String> = emptyMap(),
     val type: String = "normal",
     private val fbColor: List<Int> = emptyList(),
     private val bgColor: List<Int> = emptyList(),
 ) {
+    val text: String get() = textByLanguage["text"] ?: throw Exception("Dialog not found")
+    fun getTextBy(language: Language) = textByLanguage[language.target] ?: text
+
     val width get() = x2 - x1
     val height get() = y2 - y1
     val centerY get() = (y2 + y1) / 2f
@@ -62,14 +66,15 @@ private object DialogListSerializer :
     override fun transformDeserialize(element: JsonElement): JsonElement {
         return JsonArray(
             element.jsonArray.map { jsonElement ->
-                val (coordinates, text) = getCoordinatesAndDialog(jsonElement)
+                val coordinates = getCoordinates(jsonElement)
+                val textByLanguage = getDialogs(jsonElement)
 
                 buildJsonObject {
                     put("x1", coordinates[0])
                     put("y1", coordinates[1])
                     put("x2", coordinates[2])
                     put("y2", coordinates[3])
-                    put("text", text)
+                    put("textByLanguage", textByLanguage)
 
                     try {
                         val obj = jsonElement.jsonObject
@@ -85,13 +90,28 @@ private object DialogListSerializer :
         )
     }
 
-    private fun getCoordinatesAndDialog(element: JsonElement): Pair<JsonArray, JsonElement> {
+    private fun getCoordinates(element: JsonElement): JsonArray {
         return try {
-            val arr = element.jsonArray
-            arr[0].jsonArray to arr[1]
+            element.jsonArray[0].jsonArray
         } catch (_: Exception) {
-            val obj = element.jsonObject
-            obj["bbox"]!!.jsonArray to obj["text"]!!
+            element.jsonObject["bbox"]!!.jsonArray
+        }
+    }
+    private fun getDialogs(element: JsonElement): JsonObject {
+        return try {
+            buildJsonObject {
+                put("text", element.jsonArray[1])
+            }
+        } catch (_: Exception) {
+            buildJsonObject {
+                // There is a problem when the "angle" is processed
+                element.jsonObject.entries.forEach {
+                    if (it.key in listOf("angle", "bbox")) {
+                        return@forEach
+                    }
+                    put(it.key, it.value)
+                }
+            }
         }
     }
 }
