+++ lib-multisrc/machinetranslations/src/eu/kanade/tachiyomi/multisrc/machinetranslations/MachineTranslationsFactoryUtils.kt
@@ -1,0 +1,5 @@
+package eu.kanade.tachiyomi.multisrc.machinetranslations
+
+class MachineTranslationsFactoryUtils
+
+data class Language(val lang: String, val target: String = lang, val origin: String = "en")
